fix lint warnings + usePreviewDepth hook
